feat: add AuthInput
